% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{hcrfetch}
\alias{hcrfetch}
\title{function to fetch data directly}
\usage{
hcrfetch(
  src,
  dataset,
  file,
  path = here::here("data", dataset, file),
  cache = TRUE
)
}
\description{
function to fetch data directly
}
\author{
Hisham Galal
}
