% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo.R
\name{kobo.conn}
\alias{kobo.conn}
\title{Connect to UNHCR kobo server: http://kobo.unhcr.org}
\usage{
kobo.conn()
}
\description{
Connect to UNHCR kobo server using API key stored in your .Renviron file
}
\author{
Hisham Galal
}
