% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl.R
\name{mdl.conn}
\alias{mdl.conn}
\title{Connect to UNHCR mdl server http://microdata.unhcr.org}
\usage{
mdl.conn()
}
\description{
Connect to UNHCR mdl server using API key stored in your .Renviron file
}
\author{
Hisham Galal
}
