% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridl.R
\name{ridl.conn}
\alias{ridl.conn}
\title{Connect to UNHCR ridl server http://ridl.unhcr.org}
\usage{
ridl.conn()
}
\description{
Connect to UNHCR ridl server using API key stored in your .Renviron file
}
\author{
Hisham Galal
}
