% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcrdata-package.R
\docType{package}
\name{hcrdata-package}
\alias{hcrdata}
\alias{hcrdata-package}
\title{hcrdata: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
