% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcrdata-package.R
\docType{package}
\name{hcrdata-package}
\alias{hcrdata}
\alias{hcrdata-package}
\title{hcrdata: Uniform Interface to UNHCR Data}
\description{
Add a more detailed description here...
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/galalH/hcrdata}
  \item Report bugs at \url{https://github.com/galalH/hcrdata/issues}
}

}
\keyword{internal}
